// apps/console/src/__generated__/core/UpdateVersionDialogFragment.graphql.ts
export type DocumentStatus = "DRAFT" | "PUBLISHED";
import { FragmentRefs } from "relay-runtime";
export type UpdateVersionDialogFragment$data = {
  readonly id: string;
  readonly versions: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly content: string;
        readonly id: string;
        readonly status: DocumentStatus;
      };
    }>;
  };
  readonly " $fragmentType": "UpdateVersionDialogFragment";
};
export type UpdateVersionDialogFragment$key = {
  readonly " $data"?: UpdateVersionDialogFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"UpdateVersionDialogFragment">;
};